+++ gcip2/__init__.py
@@ -1,6 +1,5 @@
-from gcip2.pipeline_core.gitlab_ci import GitlabCiBuilderImpl
-from gcip2.pipeline_core.jobs.base import Base, BaseLinux, BaseTask
-from gcip2.pipeline_core.pipeline import PipelineBuilderImpl
+from gcip2.pipeline import GitlabCiBuilderImpl, PipelineBuilderImpl
+from gcip2.pipeline.jobs.base import Base, BaseLinux, BaseTask
 from gcip2.pipeline_vars import Predefined
 from gcip2.project_config import ProjectConfig
 from gcip2.vault import SecretsHandler