+++ gcip2/pipeline/__init__.py
@@ -1,0 +1,3 @@
+from gcip2.pipeline.pipeline import GitlabCiBuilderImpl, PipelineBuilderImpl
+
+__all__ = ("PipelineBuilderImpl", "GitlabCiBuilderImpl")