+++ pipelines/initialization/ci.py
@@ -3,6 +3,7 @@ from typing import Self
 
 from _tasks._consts import Tasks
 from gcip2 import GitlabCiBuilderImpl, PipelineBuilderImpl
+from gcip2.pipeline.jobs.base import BaseTask
 from gcip2.pipeline_core import (
     JobBuilderImpl,
     Workflow,
@@ -12,7 +13,6 @@ from gcip2.pipeline_core import (
     WorkflowRule,
     WorkflowWhen,
 )
-from gcip2.pipeline_core.jobs.base import BaseTask
 
 
 class Stages(str, Enum):