+++ pipelines/unit/ci.py
@@ -1,6 +1,7 @@
 from typing import Self
 
 from gcip2 import GitlabCiBuilderImpl, PipelineBuilderImpl
+from gcip2.pipeline.jobs.base import BaseLinux
 from gcip2.pipeline_core import (
     ArtifactsReports,
     ArtifactsReportsCoverage,
@@ -12,7 +13,6 @@ from gcip2.pipeline_core import (
     WorkflowRule,
     WorkflowWhen,
 )
-from gcip2.pipeline_core.jobs.base import BaseLinux
 
 
 class RunUnitTests(JobBuilderImpl):